+++ packages/extension-base/src/services/earning-service/handlers/native-staking/base.ts
@@ -4,6 +4,7 @@
 import { TransactionError } from '@subwallet/extension-base/background/errors/TransactionError';
 import { ChainType, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import KoniState from '@subwallet/extension-base/koni/background/handlers/State';
+import { STAKING_IDENTITY_API_SLUG } from '@subwallet/extension-base/services/earning-service/constants';
 import { BasicTxErrorType, EarningRewardHistoryItem, EarningRewardItem, HandleYieldStepData, OptimalYieldPath, OptimalYieldPathParams, RequestBondingSubmit, SubmitChangeValidatorStaking, SubmitJoinNativeStaking, SubmitYieldJoinData, TransactionData, ValidatorInfo, YieldPoolMethodInfo, YieldPoolType, YieldPositionInfo, YieldStepBaseInfo, YieldStepType, YieldTokenBaseInfo } from '@subwallet/extension-base/types';
 
 import { noop } from '@polkadot/util';
@@ -40,6 +41,7 @@ export default abstract class BaseNativeStakingPoolHandler extends BasePoolHandl
     this.slug = `${symbol}___native_staking___${_chainInfo.slug}`;
     this.name = `${_chainInfo.name} Native Staking`;
     this.shortName = _chainInfo.name.replaceAll(' Relay Chain', '');
+    this.canOverrideIdentity = !!STAKING_IDENTITY_API_SLUG[chain];
   }
 
   protected getDescription (amount = '0'): string {
@@ -160,7 +162,7 @@ export default abstract class BaseNativeStakingPoolHandler extends BasePoolHandl
 
   async handleYieldJoin (_data: SubmitYieldJoinData, path: OptimalYieldPath, currentStep: number): Promise<HandleYieldStepData> {
     const data = _data as SubmitJoinNativeStaking;
-    const { address, amount, selectedValidators, slug } = data;
+    const { address, amount, selectedValidators, slug, subnetData } = data;
 
     const positionInfo = await this.getPoolPosition(address, slug);
     const [extrinsic] = await this.createJoinExtrinsic(data, positionInfo);
@@ -170,7 +172,8 @@ export default abstract class BaseNativeStakingPoolHandler extends BasePoolHandl
       slug: this.slug,
       amount,
       address,
-      selectedValidators
+      selectedValidators,
+      subnetData
     };
 
     return {