+++ packages/extension-base/src/services/transaction-service/utils.ts
@@ -2,11 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { _ChainInfo } from '@subwallet/chain-list/types';
-import { ExtrinsicDataTypeMap, ExtrinsicsDataResponse, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
+import { ExtrinsicDataTypeMap, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { _getBlockExplorerFromChain, _isChainTestNet, _isPureBitcoinChain, _isPureCardanoChain, _isPureEvmChain, _isPureTonChain } from '@subwallet/extension-base/services/chain-service/utils';
 import { CHAIN_FLIP_MAINNET_EXPLORER, CHAIN_FLIP_TESTNET_EXPLORER, SIMPLE_SWAP_EXPLORER } from '@subwallet/extension-base/services/swap-service/utils';
 import { ChainflipSwapTxData, SimpleSwapTxData } from '@subwallet/extension-base/types/swap';
-import { SWApiResponse } from '@subwallet/subwallet-api-sdk/types';
 
 import { hexAddPrefix, isHex, u8aToHex } from '@polkadot/util';
 import { decodeAddress } from '@polkadot/util-crypto';
@@ -23,6 +22,26 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
     return 'accounts';
   }
 
+  if (explorerLink.includes('edgscan.ink')) {
+    return 'accounts';
+  }
+
+  if (explorerLink.includes('devnet-explorer.mosaicchain.io')) {
+    return 'accounts';
+  }
+
+  if (explorerLink.includes('statescan.io')) {
+    return '#/accounts';
+  }
+
+  if (explorerLink.includes('explorer.joystream.org')) {
+    return '#/accounts';
+  }
+
+  if (explorerLink.includes('explorer.gen6.app')) {
+    return '#/accounts';
+  }
+
   if (explorerLink.includes('deeperscan.io')) {
     return 'account';
   }
@@ -35,19 +54,15 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
     return 'account';
   }
 
-  if (explorerLink.includes('statescan.io')) {
-    return '#/accounts';
-  }
-
-  if (explorerLink.includes('explorer.gen6.app')) {
-    return '#/accounts';
-  }
-
-  if (explorerLink.includes('astral.autonomys')) {
-    return 'accounts';
+  if (explorerLink.includes('main.dentnet.io')) {
+    return 'account';
   }
 
   if (explorerLink.includes('taostats.io')) {
+    return 'account';
+  }
+
+  if (explorerLink.includes('uniquescan.io')) {
     return 'account';
   }
 
@@ -55,8 +70,8 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
     return '';
   }
 
-  if (explorerLink.includes('devnet-explorer.mosaicchain.io')) {
-    return 'accounts';
+  if (explorerLink.includes('pdexmon.com')) {
+    return 'holders';
   }
 
   return 'address';
@@ -64,6 +79,10 @@ function getBlockExplorerAccountRoute (explorerLink: string) {
 
 function getBlockExplorerTxRoute (chainInfo: _ChainInfo) {
   if (_isPureEvmChain(chainInfo) || _isPureBitcoinChain(chainInfo)) {
+    return 'tx';
+  }
+
+  if (['moonbeam', 'crabParachain'].includes(chainInfo.slug)) {
     return 'tx';
   }
 
@@ -75,11 +94,15 @@ function getBlockExplorerTxRoute (chainInfo: _ChainInfo) {
     return 'transaction';
   }
 
-  if (['gen6_public'].includes(chainInfo.slug)) {
+  if (['gen6_public', 'joystream'].includes(chainInfo.slug)) {
     return '#/extrinsics';
   }
 
-  if (['mosaicTest'].includes(chainInfo.slug)) {
+  if (['edgeware'].includes(chainInfo.slug)) {
+    return 'extrinsics';
+  }
+
+  if (['mosaicTest', 'polkadex'].includes(chainInfo.slug)) {
     return 'transactions';
   }
 
@@ -92,24 +115,24 @@ function getBlockExplorerTxRoute (chainInfo: _ChainInfo) {
   return 'extrinsic';
 }
 
-export function getTransactionId (value: string): Promise<string> {
-  const query = `
-    query ExtrinsicQuery {
-      extrinsics(where: {hash_eq: ${value}}, limit: 1) {
-        id
-      }
-    }`;
-
-  const apiUrl = 'https://archive-explorer.truth-network.io/graphql';
-
-  return fetch(apiUrl, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ query })
-  })
-    .then((response) => response.json())
-    .then((result: SWApiResponse<ExtrinsicsDataResponse>) => result.data.extrinsics[0].id);
-}
+// export function getTransactionId (value: string): Promise<string> {
+//   const query = `
+//     query ExtrinsicQuery {
+//       extrinsics(where: {hash_eq: ${value}}, limit: 1) {
+//         id
+//       }
+//     }`;
+//
+//   const apiUrl = 'https://archive-explorer.truth-network.io/graphql';
+//
+//   return fetch(apiUrl, {
+//     method: 'POST',
+//     headers: { 'Content-Type': 'application/json' },
+//     body: JSON.stringify({ query })
+//   })
+//     .then((response) => response.json())
+//     .then((result: SWApiResponse<ExtrinsicsDataResponse>) => result.data.extrinsics[0].id);
+// }
 
 export function getExplorerLink (chainInfo: _ChainInfo, value: string, type: 'account' | 'tx'): string | undefined {
   const explorerLink = _getBlockExplorerFromChain(chainInfo);
@@ -158,7 +181,5 @@ export function getChainflipExplorerLink (data: ChainflipSwapTxData, chainInfo: 
 }
 
 export function getSimpleSwapExplorerLink (data: SimpleSwapTxData) {
-  const simpleswapDomain = SIMPLE_SWAP_EXPLORER;
-
-  return `${simpleswapDomain}/exchange?id=${data.id}`;
+  return `${SIMPLE_SWAP_EXPLORER}/exchange?id=${data.id}`;
 }